UPDATES:
 * Finalize Python3 support (remove unicode)
 * Change Readme
 * Add "setup.py"
