+++ inflector/__init__.py
@@ -1,0 +1,7 @@
+#coding=utf-8
+
+from .rules.english import English
+from .rules.french import French
+from .rules.spanish import Spanish
+
+from .inflector import Inflector
