+++ inflector/rules/base.py
