+++ inflector/rules/english.py
