+++ inflector/rules/french.py
