+++ setup.py
@@ -1,0 +1,13 @@
+# coding = utf-8
+
+from setuptools import setup, find_packages
+
+
+setup(name='inflector',
+      version='0.1a',
+      description='Inflects English, French and Spanish nouns. Similar to Rails inflector but for Python',
+      url='https://github.com/bermi/Python-Inflector',
+      author='Bermi Ferrer',
+      license='BSD',
+      packages=['inflector'],
+      zip_safe=False)
